# Add fast path for identically named data frames in `df_ptype2()`