--- a/src/type-data-frame.c
+++ b/src/type-data-frame.c
@@ -287,46 +287,39 @@ SEXP vctrs_df_ptype2(SEXP x, SEXP y, SEXP x_arg, SEXP y_arg) {
   return df_ptype2(x, y, &x_arg_, &y_arg_);
 }
 
-static SEXP df_ptype2_impl(SEXP x, SEXP y,
-                           SEXP x_names, SEXP y_names,
-                           R_len_t x_len, R_len_t y_len,
+static SEXP df_ptype2_impl(SEXP x, SEXP y, SEXP x_names, SEXP y_names,
                            struct vctrs_arg* x_arg, struct vctrs_arg* y_arg);
 
-static SEXP df_ptype2_sequential(SEXP x, SEXP y,
-                                 SEXP names, R_len_t len,
+static SEXP df_ptype2_sequential(SEXP x, SEXP y, SEXP names,
                                  struct vctrs_arg* x_arg, struct vctrs_arg* y_arg);
-
-static bool identical_names(SEXP x_names, SEXP y_names, R_len_t x_len, R_len_t y_len);
 
 // [[ include("vctrs.h") ]]
 SEXP df_ptype2(SEXP x, SEXP y, struct vctrs_arg* x_arg, struct vctrs_arg* y_arg) {
   SEXP x_names = PROTECT(r_names(x));
   SEXP y_names = PROTECT(r_names(y));
 
-  R_len_t x_len = Rf_length(x_names);
-  R_len_t y_len = Rf_length(y_names);
-
   SEXP out;
 
-  if (identical_names(x_names, y_names, x_len, y_len)) {
-    out = df_ptype2_sequential(x, y, x_names, x_len, x_arg, y_arg);
+  if (equal_object(x_names, y_names)) {
+    out = df_ptype2_sequential(x, y, x_names, x_arg, y_arg);
   } else {
-    out = df_ptype2_impl(x, y, x_names, y_names, x_len, y_len, x_arg, y_arg);
+    out = df_ptype2_impl(x, y, x_names, y_names, x_arg, y_arg);
   }
 
   UNPROTECT(2);
   return out;
 }
 
-SEXP df_ptype2_impl(SEXP x, SEXP y,
-                    SEXP x_names, SEXP y_names,
-                    R_len_t x_len, R_len_t y_len,
+SEXP df_ptype2_impl(SEXP x, SEXP y, SEXP x_names, SEXP y_names,
                     struct vctrs_arg* x_arg, struct vctrs_arg* y_arg) {
   SEXP x_dups_pos = PROTECT(vec_match(x_names, y_names));
   SEXP y_dups_pos = PROTECT(vec_match(y_names, x_names));
 
   int* x_dups_pos_data = INTEGER(x_dups_pos);
   int* y_dups_pos_data = INTEGER(y_dups_pos);
+
+  R_len_t x_len = Rf_length(x_names);
+  R_len_t y_len = Rf_length(y_names);
 
   // Count columns that are only in `y`
   R_len_t rest_len = 0;
@@ -384,9 +377,10 @@ SEXP df_ptype2_impl(SEXP x, SEXP y,
   return out;
 }
 
-SEXP df_ptype2_sequential(SEXP x, SEXP y,
-                          SEXP names, R_len_t len,
+SEXP df_ptype2_sequential(SEXP x, SEXP y, SEXP names,
                           struct vctrs_arg* x_arg, struct vctrs_arg* y_arg) {
+  R_len_t len = Rf_length(names);
+
   SEXP out = PROTECT(Rf_allocVector(VECSXP, len));
   Rf_setAttrib(out, R_NamesSymbol, names);
 
@@ -413,30 +407,6 @@ SEXP df_ptype2_sequential(SEXP x, SEXP y,
 
   UNPROTECT(1);
   return out;
-}
-
-static bool identical_names(SEXP x_names, SEXP y_names, R_len_t x_len, R_len_t y_len) {
-  if (x_names == y_names) {
-    return true;
-  }
-
-  if (x_len != y_len) {
-    return false;
-  }
-
-  const SEXP* p_x_names = STRING_PTR_RO(x_names);
-  const SEXP* p_y_names = STRING_PTR_RO(y_names);
-
-  for (R_len_t i = 0; i < x_len; ++i) {
-    const SEXP x_name = p_x_names[i];
-    const SEXP y_name = p_y_names[i];
-
-    if (x_name != y_name) {
-      return false;
-    }
-  }
-
-  return true;
 }
 
 // [[ register() ]]