Closes #979 

This PR adds early exits for `df_ptype2()` when the names of the two data frames are either the exact same reference or have identical names. 

``` r
library(vctrs)
library(tibble)
library(rlang)
```

Empty tibble

```r
x <- tibble()

# Master
bench::mark(vec_ptype2(x, x), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(x, x)   2.22µs   2.97µs   319680.        0B     16.0

# This PR
bench::mark(vec_ptype2(x, x), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(x, x)   1.15µs   1.31µs   684663.        0B     20.5
```

1 row, 100 cols, identical names

```r
y <- as_tibble_row(set_names(as.list(1:100), paste0("a", 1:100)))

# Master
bench::mark(vec_ptype2(y, y), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(y, y)     11µs   13.9µs    66197.    6.42KB     7.94

# This PR
bench::mark(vec_ptype2(y, y), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(y, y)   5.11µs   5.58µs   162741.      848B     4.88
```

Same as before, but different objects with the same names so we have to loop through the names to check that they are identical.

```r
z1 <- as_tibble_row(set_names(as.list(1:100), paste0("a", 1:100)))
z2 <- as_tibble_row(set_names(as.list(1:100), paste0("a", 1:100)))

# Master
bench::mark(vec_ptype2(z1, z2))
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(z1, z2)   11.2µs   13.4µs    71764.    6.42KB     7.18

# This PR
bench::mark(vec_ptype2(z1, z2))
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(z1, z2)   6.91µs   7.74µs   120376.      848B     12.0
```

The last name is different here, everything else is the same. This is the worst case since we have to iterate through all the names and then fall back to matching.

```r
w1 <- as_tibble_row(set_names(as.list(1:100), paste0("a", 1:100)))
w2 <- as_tibble_row(set_names(as.list(1:100), c(paste0("a", 1:99), "a101")))

# Master
bench::mark(vec_ptype2(w1, w2))
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(w1, w2)   11.4µs   13.3µs    73539.    6.44KB     7.35

# This PR
bench::mark(vec_ptype2(w1, w2))
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(w1, w2)   11.6µs   16.5µs    58799.    6.44KB     5.88
```

Computing the common type of many data frames with identical names (like from `vec_chop()`)

```r
df <- tibble(x = 1:10000, y = 1:10000)
dfs <- vec_chop(df)

ptype_common <- function(x) {
  vec_ptype_common(!!!x)
}

# Master
bench::mark(ptype_common(dfs))
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 ptype_common(dfs)   19.9ms   23.4ms      42.6    6.92KB     47.3

# This PR
bench::mark(ptype_common(dfs))
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 ptype_common(dfs)   4.12ms   5.66ms      174.    6.92KB     16.2
```