@DavisVaughan 2020-04-06
Nice performance boost for `vec_rbind()`. I imagine it is fairly common to rbind data frames with the same column names, even if the column names aren't identical references. Like this:

``` r
library(vctrs)
library(tibble)
#> Warning: package 'tibble' was built under R version 3.6.2

x <- vector("list", 10000)

# ensure the column names aren't created from the same reference
for (i in seq_along(x)) {
  x[[i]] <- tibble(x = 1, y = 2)
}

vecrbind <- function(x) {
  vec_rbind(!!!x)
}

# master
bench::mark(vecrbind(x), iterations = 50)
#> # A tibble: 1 x 6
#>   expression       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vecrbind(x)   80.1ms   80.1ms      12.5     278KB     724.

# this pr
bench::mark(vecrbind(x), iterations = 50)
#> # A tibble: 1 x 6
#>   expression       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vecrbind(x)   40.2ms   48.9ms      20.1     278KB     32.9
```